feat: 타입을 정의하는 declare module 부분을 types 폴더로 따로 분리 하여 사용 할 수 있도록 작업
