+++ src/common/theme.tsx
@@ -1,29 +1,7 @@
 import { createTheme } from '@mui/material/styles';
-import React from 'react';
 
 import color from '@constants/color';
 import fontSize from '@constants/fontSize';
-
-declare module '@mui/material/styles' {
-  interface TypographyVariants {
-    input1: React.CSSProperties;
-    input2: React.CSSProperties;
-  }
-
-  // allow configuration using `createTheme`
-  interface TypographyVariantsOptions {
-    input1?: React.CSSProperties;
-    input2?: React.CSSProperties;
-  }
-}
-
-// Update the Typography's variant prop options
-declare module '@mui/material/Typography' {
-  interface TypographyPropsVariantOverrides {
-    input1: true;
-    input2: true;
-  }
-}
 
 export const theme = createTheme({
   palette: {
